ansible: install and test the hermes driver instead of xdma
This Ansible playbook is called by the eid-hermes-qemu Docker[1] to
when setting up a Hermes-enabled VM.

The XDMA driver is no longer being used, so we change it to install the
Hermes driver and add tests to run the unit tests and examples.

In the future, we plan to use it for continuous integration of
eid-hermes and eid-hermes-qemu.

[1]: https://github.com/Eideticom/eid-hermes-qemu/tree/hermes/docker/hermes